+++ apm-sniffer/apm-sdk-plugin/elasticsearch-6.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/elasticsearch/v6/define/IndicesClientInstrumentation.java
@@ -68,7 +68,7 @@ public class IndicesClientInstrumentation extends ClassEnhancePluginDefine {
 
                 @Override
                 public boolean isOverrideArgs() {
-                    return false;
+                    return true;
                 }
             },
             new InstanceMethodsInterceptPoint() {
@@ -85,7 +85,7 @@ public class IndicesClientInstrumentation extends ClassEnhancePluginDefine {
 
                 @Override
                 public boolean isOverrideArgs() {
-                    return false;
+                    return true;
                 }
             },
             new InstanceMethodsInterceptPoint() {
@@ -102,7 +102,7 @@ public class IndicesClientInstrumentation extends ClassEnhancePluginDefine {
 
                 @Override
                 public boolean isOverrideArgs() {
-                    return false;
+                    return true;
                 }
             },
                 new InstanceMethodsInterceptPoint() {
@@ -119,7 +119,7 @@ public class IndicesClientInstrumentation extends ClassEnhancePluginDefine {
 
                     @Override
                     public boolean isOverrideArgs() {
-                        return false;
+                        return true;
                     }
                 }
         };