+++ apm-sniffer/apm-sdk-plugin/elasticsearch-6.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/elasticsearch/v6/interceptor/IndicesClientAnalyzeMethodsInterceptor.java
@@ -30,6 +30,8 @@ import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.InstanceM
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.MethodInterceptResult;
 import org.apache.skywalking.apm.network.trace.component.ComponentsDefine;
 import org.apache.skywalking.apm.plugin.elasticsearch.common.RestClientEnhanceInfo;
+import org.apache.skywalking.apm.plugin.elasticsearch.v6.support.AdapterUtil;
+import org.elasticsearch.action.ActionListener;
 import org.elasticsearch.action.admin.indices.analyze.AnalyzeRequest;
 
 import java.lang.reflect.Method;
@@ -58,6 +60,10 @@ public class IndicesClientAnalyzeMethodsInterceptor implements InstanceMethodsAr
                 Tags.DB_STATEMENT.set(span, analyzeRequest.text()[0]);
             }
             SpanLayer.asDB(span);
+            if (allArguments.length > 2 && allArguments[2] != null && allArguments[2] instanceof ActionListener) {
+                allArguments[2] = AdapterUtil.wrapActionListener(restClientEnhanceInfo, Constants.ANALYZE_OPERATOR_NAME,
+                                                                 (ActionListener) allArguments[2]);
+            }
         }
     }