+++ apm-sniffer/apm-sdk-plugin/elasticsearch-7.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/elasticsearch/v7/define/IndicesClientInstrumentation.java
@@ -62,7 +62,7 @@ public class IndicesClientInstrumentation extends ClassEnhancePluginDefine {
 
                 @Override
                 public boolean isOverrideArgs() {
-                    return false;
+                    return true;
                 }
             },
             new InstanceMethodsInterceptPoint() {
@@ -79,7 +79,7 @@ public class IndicesClientInstrumentation extends ClassEnhancePluginDefine {
 
                 @Override
                 public boolean isOverrideArgs() {
-                    return false;
+                    return true;
                 }
             },
             new InstanceMethodsInterceptPoint() {
@@ -96,7 +96,7 @@ public class IndicesClientInstrumentation extends ClassEnhancePluginDefine {
 
                 @Override
                 public boolean isOverrideArgs() {
-                    return false;
+                    return true;
                 }
             },
             new InstanceMethodsInterceptPoint() {
@@ -113,7 +113,7 @@ public class IndicesClientInstrumentation extends ClassEnhancePluginDefine {
 
                 @Override
                 public boolean isOverrideArgs() {
-                    return false;
+                    return true;
                 }
             }
         };