+++ apm-sniffer/apm-sdk-plugin/elasticsearch-7.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/elasticsearch/v7/interceptor/IndicesClientAnalyzeMethodsInterceptor.java
@@ -31,6 +31,8 @@ import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.MethodInt
 import org.apache.skywalking.apm.network.trace.component.ComponentsDefine;
 import org.apache.skywalking.apm.plugin.elasticsearch.common.RestClientEnhanceInfo;
 import org.apache.skywalking.apm.plugin.elasticsearch.v7.Constants;
+import org.apache.skywalking.apm.plugin.elasticsearch.v7.support.AdapterUtil;
+import org.elasticsearch.action.ActionListener;
 import org.elasticsearch.client.indices.AnalyzeRequest;
 
 import java.lang.reflect.Method;
@@ -59,6 +61,10 @@ public class IndicesClientAnalyzeMethodsInterceptor implements InstanceMethodsAr
                 Tags.DB_STATEMENT.set(span, analyzeRequest.text()[0]);
             }
             SpanLayer.asDB(span);
+            if (allArguments.length > 2 && allArguments[2] != null && allArguments[2] instanceof ActionListener) {
+                allArguments[2] = AdapterUtil.wrapActionListener(restClientEnhanceInfo, Constants.ANALYZE_OPERATOR_NAME,
+                                                                 (ActionListener) allArguments[2]);
+            }
         }
     }